requirements.txt
openai-whisper
torch
torchaudio
numpy
sounddevice
scipy
librosa
webrtcvad
pyaudio 